B1: introduce DatabaseContext; refactor AppState to a map
Splits all per-database state (SQLite connection, directory tree
cache, file-existence cache, refresh coordination) out of AppState
into a new DatabaseContext struct. AppState now owns
`databases: HashMap<slug, Arc<DatabaseContext>>` plus the global
pregeneration config and the cache root path.

For B1, AppState always has exactly one entry, and retains all its
public field/method shape via compatibility shims (Arc clones of the
single context's caches, plus delegating wrappers around db(),
get_directory_tree(), ensure_cache_available(), etc.). This keeps
~45 handler call sites untouched until B2 introduces the path-nested
API and DbContext extractor.

Also adds:
- `server::slug` module with `compute_default_slug` (FNV-1a 32-bit
  over the canonicalized DB path) and `validate_slug` (URL-safe rules).
  These are the seed for default slugs and the validator for
  user-edited ones.
- `single_context()` helper on AppState that panics if there's not
  exactly one DB — explicit marker for code B2 must convert.

Build clean with and without the tauri feature. cargo test passes
(13 integration tests + 5 new slug unit tests). cargo clippy clean.
